Refactor settings module
