--- djiometsa/dev_settings.py
@@ -1,4 +1,0 @@
-
-DEBUG = True
-
-ALLOWED_HOSTS = ['*']
