+++ djiometsa/settings/dev.py
@@ -1,0 +1,11 @@
+from .base import *
+
+DEBUG = True
+
+CSRF_COOKIE_SECURE = False
+
+SESSION_COOKIE_SECURE = False
+
+SECURE_SSL_REDIRECT = False
+
+ALLOWED_HOSTS = ['*']
