--- resume/models.py
@@ -1,3 +1,0 @@
-from django.db import models
-# from django_mysql.models import ListCharField, JSONField
-# Create your models here.
